+++ docs/source/releases.rst
@@ -21,6 +21,7 @@ Additions
 - Added :meth:`~BaseMessageChannel.messages` method to iterate through channels history.
 - Added :meth:`Guild.members` method to iterate through guild members.
 - Added :attr:`PrivateChannel.url`, :attr:`GuildChannel.url` and :attr:`Message.url` properties
+- Added :meth:`BaseMessageChannel.trigger_typing` and :meth:`BaseMessageChannel.typing` for working with typing indicators.
 
 Changes
 ~~~~~~~~